Chore: Improvements and visual bug fixes in the "SegmenteControl" component ♻️
This component was having a lot of visual bugs where its indicator did not follow the component as a whole properly, especially in terms of responsiveness. By using "ResizeObserve" I was able to fix this. I also made a small improvement and added separators between the element options, which can also be disabled.